Drop support of 2020.1 platform
