--- clion/src/201/main/kotlin/org/rust/clion/profiler/compatUtils.kt
@@ -1,11 +1,0 @@
-/*
- * Use of this source code is governed by the MIT license that can be
- * found in the LICENSE file.
- */
-
-package org.rust.clion.profiler
-
-import com.intellij.util.containers.WeakStringInterner
-
-typealias NativeCallStackElementRenderer = com.intellij.profiler.ui.flamechart.NativeCallChartNodeRenderer
-typealias StringInterner = WeakStringInterner
