--- clion/src/202/main/kotlin/org/rust/clion/profiler/compatUtils.kt
@@ -1,12 +1,0 @@
-/*
- * Use of this source code is governed by the MIT license that can be
- * found in the LICENSE file.
- */
-
-package org.rust.clion.profiler
-
-import com.intellij.util.containers.Interner
-
-// BACKCOMPAT: 2020.1. Inline it
-typealias NativeCallStackElementRenderer = com.intellij.profiler.ui.NativeCallStackElementRenderer
-typealias StringInterner = Interner<String>
