--- debugger/src/201/main/kotlin/org/rust/debugger/lang/RsDebuggerEditorsExtension.kt
@@ -1,12 +1,0 @@
-/*
- * Use of this source code is governed by the MIT license that can be
- * found in the LICENSE file.
- */
-
-package org.rust.debugger.lang
-
-import org.rust.lang.RsLanguage
-
-class RsDebuggerEditorsExtension : RsDebuggerEditorsExtensionBase() {
-    override fun getSupportedLanguage() = RsLanguage
-}
