--- debugger/src/201/main/resources/META-INF/platform-core-debugger-only.xml
@@ -1,5 +1,0 @@
-<idea-plugin>
-    <extensions defaultExtensionNs="cidr.debugger">
-        <editorsExtension implementation="org.rust.debugger.lang.RsDebuggerEditorsExtension"/>
-    </extensions>
-</idea-plugin>
