--- debugger/src/202/main/kotlin/org/rust/debugger/lang/RsDebuggerEditorsExtension.kt
@@ -1,8 +1,0 @@
-/*
- * Use of this source code is governed by the MIT license that can be
- * found in the LICENSE file.
- */
-
-package org.rust.debugger.lang
-
-class RsDebuggerEditorsExtension : RsDebuggerEditorsExtensionBase()
