+++ debugger/src/202/main/resources/META-INF/platform-core-debugger-only.xml
@@ -1,5 +1,2 @@
 <idea-plugin>
-    <extensions defaultExtensionNs="cidr.debugger">
-        <editorsExtension language="Rust" implementationClass="org.rust.debugger.lang.RsDebuggerEditorsExtension"/>
-    </extensions>
 </idea-plugin>
