--- gradle-201.properties
@@ -1,13 +1,0 @@
-ideaVersion=IU-2020.1.2
-clionVersion=CL-2020.1.2
-
-# https://plugins.jetbrains.com/plugin/12775-native-debugging-support/versions
-nativeDebugPluginVersion=201.7223.18
-# https://plugins.jetbrains.com/plugin/12175-grazie/versions
-graziePluginVersion=201.7846.2
-# https://plugins.jetbrains.com/plugin/227-psiviewer/versions
-psiViewerPluginVersion=201.6251.22-EAP-SNAPSHOT.3
-
-# please see http://www.jetbrains.org/intellij/sdk/docs/basics/getting_started/build_number_ranges.html for description
-sinceBuild=201.6487
-untilBuild=201.*
