+++ gradle.properties
@@ -1,5 +1,5 @@
 propertiesPluginEnvironmentNameProperty=platformVersion
-# Supported platforms: 201, 202
+# Supported platforms: 202
 platformVersion=202
 # Supported IDEs: idea, clion
 baseIDE=idea
