--- src/201/main/resources/META-INF/platform-core.xml
@@ -1,2 +1,0 @@
-<idea-plugin>
-</idea-plugin>
