+++ src/main/kotlin/org/rust/ide/lineMarkers/RsRecursiveCallLineMarkerProvider.kt
@@ -25,7 +25,7 @@ class RsRecursiveCallLineMarkerProvider : LineMarkerProvider {
 
     override fun getLineMarkerInfo(element: PsiElement): LineMarkerInfo<*>? = null
 
-    override fun collectSlowLineMarkers(elements: List<PsiElement>, result: SlowRunMarketResult) {
+    override fun collectSlowLineMarkers(elements: List<PsiElement>, result: MutableCollection<in LineMarkerInfo<*>>) {
         val lines = HashSet<Int>()  // To prevent several markers on one line
 
         for (el in elements) {
