+++ src/main/resources/META-INF/rust-core.xml
@@ -1014,6 +1014,8 @@
     <applicationListeners>
         <listener class="org.rust.ide.navigation.goto.RsGoToDeclarationRunningService$Listener"
                   topic="com.intellij.openapi.actionSystem.ex.AnActionListener" />
+        <listener class="org.rust.ide.dynamic.RsDynamicPluginListener"
+                  topic="com.intellij.ide.plugins.DynamicPluginListener" />
     </applicationListeners>
 
     <projectListeners>
